reorganise code
